+++ src/datasets/table.py
@@ -1854,6 +1854,7 @@ def array_concat(arrays: List[pa.Array]):
         return concatenated_offsets
 
     def _concat_arrays(arrays):
+        array_type = arrays[0].type
         if isinstance(array_type, pa.PyExtensionType):
             return array_type.wrap_array(_concat_arrays([array.storage for array in arrays]))
         elif pa.types.is_struct(array_type):
